+++ packages/app-bridge/src/registries/api/GetCurrentUser.ts
@@ -1,0 +1,10 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+export type GetCurrentUserPayload = void;
+
+export type GetCurrentUserResponse = {
+    id: string;
+    name?: string | null;
+    avatar?: string | null;
+    email: string;
+};