+++ packages/app-bridge/src/registries/commands/index.ts
@@ -2,6 +2,7 @@
 
 export * from './AssetChooser';
 export * from './AssetViewer';
+export * from './CommandRegistry';
 export * from './NavigationManager';
 export * from './TemplateChooser';
 export * from './CreateNewPublication';