+++ packages/sidebar-settings/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @frontify/sidebar-settings
 
+## 0.6.3
+
+### Patch Changes
+
+- [#502](https://github.com/Frontify/brand-sdk/pull/502) [`d1a60e9`](https://github.com/Frontify/brand-sdk/commit/d1a60e9fabb2920d3fc716e18f58d8de1527ab26) Thanks [@fulopdaniel](https://github.com/fulopdaniel)! - feat(\*): add link block
+
+## 0.6.2
+
+### Patch Changes
+
+- [`22684e3`](https://github.com/Frontify/brand-sdk/commit/22684e358fd2b073790a0959606c6a5d44e35c4e) Thanks [@SamuelAlev](https://github.com/SamuelAlev)! - Update dependencies
+
 ## 0.6.1
 
 ### Patch Changes